Minimise checkout size for building the book
Since building the book doesn't require any git history in its current
form, we can minimise the amount of network, disk and processing required.

This makes the most difference when the repo contains a lot of data in
its history, e.g. when many binary blobs like images have been replaced
over time, or when the book is part of a larger repo.